+++ liquid/golden/__init__.py
@@ -77,6 +77,7 @@ from . import base64_url_safe_encode_filter
 from . import base64_url_safe_decode_filter
 
 from . import range_objects
+from . import special
 
 test_cases = [
     assign_tag,
@@ -153,4 +154,5 @@ test_cases = [
     base64_url_safe_encode_filter,
     base64_url_safe_decode_filter,
     range_objects,
+    special,
 ]